Prerequisites for app building and logging
Added the requirements and changed the app so it will work if built as an exe.
The app building process is based on pyinstaller and can be done either manually on a local machine or via docker+wine.
Also logging was added for better troubleshooting in the process. Updates in the wiki will follow if changes are accepted.