feature: Powershell script for installation
Due to the fact that an automatic deployment of the msi via pipeline proves to be too difficult,
a powershell script was written which moves all important files into the installation directory,
creates a Windows Defender rule, sets the directory permissions and creates a shortcut on the desktop of the user.
If the directory exists, an uninstall routine is executed.

Resolves: #95